requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.350
faiss-cpu==1.7.4
python-dotenv==1.0.0
openai==1.3.5
pydantic==2.5.2
